Fix MainActivity interface conflicts and math function errors

+++ AndroidApp/src/main/java/com/multisensor/recording/controllers/CalibrationController.kt
@@ -989,7 +989,7 @@ class CalibrationController @Inject constructor(
         if (std == 0.0f || values.size < 4) return 3.0f
         val n = values.size
         val kurt = values.sumOf { ((it - mean) / std).pow(4).toDouble() }.toFloat()
-        return kurt * n * (n + 1) / ((n - 1) * (n - 2) * (n - 3)) - 3.0f * (n - 1).pow(2) / ((n - 2) * (n - 3))
+        return kurt * n * (n + 1) / ((n - 1) * (n - 2) * (n - 3)) - 3.0f * (n - 1).toDouble().pow(2).toFloat() / ((n - 2) * (n - 3))
     }
     
     private fun percentile(sortedValues: List<Float>, percentile: Int): Float {
